requirements.txt
Django==1.8.2
django-bootstrap3==5.4.0
django-filter==0.10.0
django-nose==1.4
django-registration-redux==1.2
django-s3direct==0.3.11
djangorestframework==3.1.2
drf-nested-routers==0.9.0
ipdb==0.8
ipython==3.1.0
Markdown==2.6.2
model-mommy==1.2.4
nose==1.3.6
psycopg2==2.6
six==1.9.0
wheel==0.24.0
